update action menu

+++ web-client/frontend/rounds/round-store.ts
@@ -1,15 +1,14 @@
 import {makeAutoObservable, observable, runInAction} from 'mobx';
 import WorkoutRoundTo from "Frontend/generated/ru/soft/common/to/WorkoutRoundTo";
 import {WorkoutRoundEndpoint} from "Frontend/generated/endpoints";
-import {uiStore} from "Frontend/common/stores/app-store";
+import {roundStore, uiStore} from "Frontend/common/stores/app-store";
 import WorkoutRoundToModel from "Frontend/generated/ru/soft/common/to/WorkoutRoundToModel";
-import WorkoutStationSnapshot from "Frontend/generated/ru/soft/common/data/snapshot/WorkoutStationSnapshot";
+import {EndpointError} from "@hilla/frontend";
 
 export class RoundStore {
     data: WorkoutRoundTo[] = [];
     filterText = '';
     selected: WorkoutRoundTo | null = null;
-    detailsOpenedStations: Array<WorkoutStationSnapshot | undefined> = [];
 
     constructor() {
         makeAutoObservable(
@@ -39,6 +38,10 @@ export class RoundStore {
         );
     }
 
+    hasSelected(): boolean {
+        return this.selected != null;
+    }
+
     updateFilter(filterText: string) {
         this.filterText = filterText;
     }
@@ -55,29 +58,55 @@ export class RoundStore {
         this.selected = selected;
     }
 
-    setDetailsOpenedStations(detailsOpenedStations: Array<WorkoutStationSnapshot | undefined>) {
-        this.detailsOpenedStations = detailsOpenedStations;
-    }
-
-    public async update(updatable: WorkoutRoundTo): Promise<void> {
+    public async update(updatable: WorkoutRoundTo) {
         if (!updatable.id) return;
-        return await WorkoutRoundEndpoint.update(updatable).then(() => {
-            this.saveLocal(updatable);
-            uiStore.showSuccess('Exercise updated.');
-        });
+        await WorkoutRoundEndpoint.update(updatable)
+            .then(() => {
+                this.saveLocal(updatable);
+                uiStore.showSuccess('Exercise updated.');
+            })
+            .catch(this.processErr);
     }
 
-    public async add(stored: WorkoutRoundTo): Promise<void> {
-        return await WorkoutRoundEndpoint.add(stored)
+    public async add(stored: WorkoutRoundTo) {
+        await WorkoutRoundEndpoint.add(stored)
             .then(stored => {
                 this.saveLocal(stored);
-                uiStore.showSuccess('Exercise created.');
-            });
+                uiStore.showSuccess('Round created.');
+            })
+            .catch(this.processErr);
+    }
+
+    public async copy() {
+        const original = roundStore.selected;
+        if (!original) return;
+
+        let copy = WorkoutRoundToModel.createEmptyValue();
+        copy.title = original.title + ' Copy ' + this.randomString(5);
+        copy.description = original.description;
+        copy.roundSchema = JSON.parse(JSON.stringify(original.roundSchema));
+
+        await WorkoutRoundEndpoint.add(copy)
+            .then(copy => {
+                this.saveLocalAfterSelected(copy);
+                uiStore.showSuccess('Round copy.');
+            })
+            .catch(this.processErr);
+    }
+
+    private randomString(length: number) {
+        const chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()';
+        const charLength = chars.length;
+        let result = '';
+        for (let i = 0; i < length; i++) {
+            result += chars.charAt(Math.floor(Math.random() * charLength));
+        }
+        return result;
     }
 
     private saveLocal(saved: WorkoutRoundTo) {
-        const contactExists = this.data.some((c) => c.id === saved.id);
-        if (contactExists) {
+        const roundExists = this.data.some((c) => c.id === saved.id);
+        if (roundExists) {
             this.data = this.data.map((existing) => {
                 if (existing.id === saved.id) {
                     return saved;
@@ -90,20 +119,40 @@ export class RoundStore {
         }
     }
 
-    async delete(removed: WorkoutRoundTo) {
-        if (!removed.id) return;
+    private saveLocalAfterSelected(copy: WorkoutRoundTo) {
+        const selected = roundStore.selected;
+        if (!selected) return;
 
-        try {
-            await WorkoutRoundEndpoint.delete(removed.id);
-            this.deleteLocal(removed);
-            uiStore.showSuccess('Contact deleted.');
-        } catch (error: any) {
-            console.log('Contact delete failed');
-            uiStore.showError(`Server error. ${error.message}`);
+        const originalExists = this.data.some((c) => c.id === selected.id);
+        if (originalExists) {
+            const dropIndex = this.data.indexOf(selected) + 1;
+            this.data.splice(dropIndex, 0, copy);
         }
+    }
+
+    async delete() {
+        const removed = roundStore.selected;
+        if (!removed || !removed.id) return;
+
+        await WorkoutRoundEndpoint.delete(removed.id)
+            .then(() => {
+                this.deleteLocal(removed);
+                uiStore.showSuccess('Round deleted.');
+                this.setSelected(null);
+            })
+            .catch(this.processErr);
     }
 
     private deleteLocal(removed: WorkoutRoundTo) {
         this.data = this.data.filter((c) => c.id !== removed.id);
+    }
+
+    protected processErr(err: any) {
+        console.log('Operation failed');
+        if (err instanceof EndpointError) {
+            uiStore.showError(`Server error. ${err.message}`);
+        } else {
+            throw err;
+        }
     }
 }
